backend/message-analyser-service/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.24.0
firebase-admin==6.0.0
python-dotenv==1.0.0
requests==2.31.0
groq==0.26.0
